requirements.txt
pandas
numpy
scikit-learn
joblib
nltk
emoji
regex
python-dotenv
pydantic

fastapi
uvicorn
streamlit
requests

tweepy
snscrape

mlflow

pytest
pyflakes
